# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API REST para sistema de comercio electrónico con autenticación JWT y pagos con Stripe",
        "contact": {
            "name": "API Support",
            "email": "support@ecommerce-api.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Renovar token JWT",
                "description": "Renueva el token JWT actual y retorna un nuevo token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Token renovado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token renovado exitosamente."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGci..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/payments/create-intent/{order}": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Crear PaymentIntent",
                "description": "Crea un PaymentIntent de Stripe para una orden específica",
                "operationId": "createPaymentIntent",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PaymentIntent creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PaymentIntent creado exitosamente."
                                        },
                                        "client_secret": {
                                            "type": "string",
                                            "example": "pi_xxx_secret_xxx"
                                        },
                                        "payment_intent_id": {
                                            "type": "string",
                                            "example": "pi_xxx"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 2849.97
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "usd"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Orden no puede ser pagada"
                    },
                    "500": {
                        "description": "Error al procesar el pago"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payments/confirm": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Confirmar pago",
                "description": "Confirma el estado de un pago en Stripe",
                "operationId": "confirmPayment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_intent_id"
                                ],
                                "properties": {
                                    "payment_intent_id": {
                                        "type": "string",
                                        "example": "pi_xxx"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pago procesado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pago procesado exitosamente."
                                        },
                                        "payment": {
                                            "$ref": "#/components/schemas/Payment"
                                        },
                                        "order": {
                                            "$ref": "#/components/schemas/Order"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "succeeded"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Pago no encontrado"
                    },
                    "500": {
                        "description": "Error al confirmar el pago"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sku": {
                        "type": "string",
                        "example": "LAPTOP-001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Laptop Gaming Pro"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1299.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 25
                    },
                    "image_url": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1299.99
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 2599.98
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "stripe_payment_intent_id": {
                        "type": "string",
                        "example": "pi_xxx"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 2849.97
                    },
                    "currency": {
                        "type": "string",
                        "example": "usd"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "succeeded",
                            "failed",
                            "refunded"
                        ]
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "card",
                        "nullable": true
                    },
                    "last4": {
                        "type": "string",
                        "example": "4242",
                        "nullable": true
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 2849.97
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "paid",
                            "cancelled",
                            "failed"
                        ]
                    },
                    "shipping_address": {
                        "type": "string",
                        "example": "Calle Principal 123, Ciudad"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "payment": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Payment"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}